fix: make the history page work with lots of events
Adds a pagination of 100 elements and a load more button to display the next batch

(cherry picked from commit 766d28e997457d7bd083b89c263b564d8a438240)
(cherry picked from commit 7178d99a77a78f9f487789525a8b3b132b3b7047)
(cherry picked from commit 0e434f3e0d69f373e6639d0a979ea46664f4459b)

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/v4/impl/ApiStateServiceImpl.java
@@ -375,8 +375,8 @@ public class ApiStateServiceImpl implements ApiStateService {
                 executionContext,
                 Arrays.asList(PUBLISH_API, EventType.UNPUBLISH_API),
                 properties,
-                0,
-                0,
+                (long) 0,
+                (long) 0,
                 0,
                 1,
                 singletonList(executionContext.getEnvironmentId())
